Changes in calderon_ref.m. Mostly added comments connecting the code to the article in sections "Andriulli method via orthogonal projections", "Definition of operators", "Methods using operators and condition numbers" and  "Norms of the error vectors in JM and Andriulli methods: V1/V2".
Notice that the numbering of operators in the code does not agree the numbering used in the article. The section "Norms of the error vectors in JM and Andriulli methods: V2" has not been included in the article because it is an equivalent approach to the problem but reducing calculations. No additional conceptual work has to be put to understand this approach once version 1 has been understood.